enemy class knockback uses floats Mapmanager created but empty

+++ BankShot/Enemy.cs
@@ -9,18 +9,20 @@ namespace BankShot
 {
     public delegate void enemyMethods(Enemy sender);
     /// <summary>
-    /// generic enemy class, handles movement, dealing damage, and death
+    /// generic enemy class, handles movement, dealing damage, and death.
+    /// Coded by Machi
     /// </summary>
     public class Enemy : Character, IDamages
     {
         //Fields
         protected int attackPower;
-        protected int knockbackDistance;
+        protected float knockbackDistance;
         public event enemyMethods enemyDeath;
         //Enemy Stats will be included as Fields
 
         //Constructor
-        public Enemy(Texture2D texture, Rectangle rect, List<Rectangle> collisionBoxes, bool active, int maxHealth, Vector2 velocity, int attackPower, int knockbackDistance)
+        public Enemy(Texture2D texture, Rectangle rect, List<Rectangle> collisionBoxes, bool active, 
+            int maxHealth, Vector2 velocity, int attackPower, float knockbackDistance)
             : base(texture,rect,collisionBoxes,active,maxHealth,velocity)
         {
             this.attackPower = attackPower;
